Major Refactoring and code revisions
Changed the Controller class to be responsible for starting the game in the correct mode depending on specific values. It'll be an adapter between the Driver class and the rest of the system.

Controller class old behaviour has been split into the editor and tester classes for higher cohesion

Tester class now uses strategy design pattern. Created package "testerstrategy" with classes "TesterStrategy", "MapFolderTesterStrategy" and "SingleMapTesterStraregy".